+++ saleor/core/middleware.py
@@ -130,7 +130,6 @@ class ReadOnlyMiddleware:
         "checkoutPaymentCreate",
         "checkoutShippingAddressUpdate",
         "checkoutShippingMethodUpdate",
-        "checkoutUpdateVoucher",
         "checkoutUpdateMetadata",
         "checkoutClearMetadata",
         "checkoutUpdatePrivateMetadata",