+++ saleor/core/permissions.py
@@ -31,6 +31,10 @@ class MenuPermissions(BasePermissionEnum):
     MANAGE_MENUS = "menu.manage_menus"
 
 
+class CheckoutPermissions(BasePermissionEnum):
+    MANAGE_CHECKOUTS = "checkout.manage_checkouts"
+
+
 class OrderPermissions(BasePermissionEnum):
     MANAGE_ORDERS = "order.manage_orders"
 
@@ -68,6 +72,7 @@ PERMISSIONS_ENUMS = [
     ShippingPermissions,
     SitePermissions,
     WebhookPermissions,
+    CheckoutPermissions,
 ]