+++ saleor/graphql/account/mutations/base.py
@@ -308,13 +308,6 @@ class CustomerInput(UserInput, UserAddressInput):
 
 
 class UserCreateInput(CustomerInput):
-    send_password_email = graphene.Boolean(
-        description=(
-            "DEPRECATED: Will be removed in Saleor 2.10, if mutation has `redirect_url`"
-            " in input then customer get email with link to set a password. "
-            "Send an email with a link to set a password."
-        )
-    )
     redirect_url = graphene.String(
         description=(
             "URL of a view where users should be redirected to "
@@ -355,19 +348,6 @@ class BaseCustomerCreate(ModelMutation, I18nMixin):
                 info=info,
             )
             cleaned_input[BILLING_ADDRESS_FIELD] = billing_address
-
-        # DEPRECATED: We should remove this condition when dropping
-        # `send_password_email` from mutation input.
-        if cleaned_input.get("send_password_email"):
-            if not cleaned_input.get("redirect_url"):
-                raise ValidationError(
-                    {
-                        "redirect_url": ValidationError(
-                            "Redirect url is required to send a password.",
-                            code=AccountErrorCode.REQUIRED,
-                        )
-                    }
-                )
 
         if cleaned_input.get("redirect_url"):
             try: